+++ src/main/java/com/example/Warterlooya/user/dto/RequestRegisterDTO.java
@@ -7,7 +7,7 @@ public class RequestRegisterDTO {
 
     private String gender;
 
-    private Long weight;
+    private int weight;
 
     private boolean oldAge;
